@@ -1,26 +1,30 @@
 {
   "name": "frontEndCapstone",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "server/index.js",
   "license": "MIT",
   "dependencies": {
     "@testing-library/react": "^13.2.0",
     "@types/jest": "^27.5.1",
     "axios": "^0.27.2",
     "eslint-plugin-babel": "^5.3.1",
+    "express": "^4.18.1",
     "jest": "^28.1.0",
     "jest-dom": "^4.0.0",
     "normalize.css": "^8.0.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-modal": "^3.15.1",
-    "react-test-renderer": "^18.1.0"
+    "react-test-renderer": "^18.1.0",
+    "webpack": "^5.73.0",
+    "webpack-merge": "^5.8.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.12",
+    "@babel/core": "^7.18.2",
     "@babel/eslint-parser": "^7.17.0",
-    "@babel/preset-env": "^7.17.12",
+    "@babel/preset-env": "^7.18.2",
     "@babel/preset-react": "^7.17.12",
+    "babel-jest": "^28.1.0",
     "babel-loader": "^8.2.5",
     "css-loader": "^6.7.1",
     "dotenv-webpack": "^7.1.0",
@@ -31,14 +35,19 @@
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "style-loader": "^3.3.1",
-    "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.9.0"
   },
   "scripts": {
-    "start": "webpack --mode=development",
-    "build": "webpack --mode=production",
-    "dev": "webpack serve --open",
-    "test": "jest --watchAll --verbose --coverage"
+    "build:prod": "webpack  -c webpack.prod.config.js",
+    "build:dev": "webpack  -c webpack.dev.config.js --mode=development -w" ,
+    "start:server": "nodemon server/index.js",
+    "test": "jest"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+    }
   }
 }
