@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import React from 'react';
 
-const apiRequest = require('./apiRequests');
+import { useInteractions } from './apiHooks';
 
 class Tracker extends React.Component {
   constructor(props) {
@@ -14,7 +14,7 @@ class Tracker extends React.Component {
     const element = event.target.className;
     const widget = child.props.className;
     const time = event.timeStamp;
-    apiRequest.postInteractions(element, widget, time)
+    useInteractions(element, widget, time)
       .then((response) => console.log(response))
       .catch((error) => console.log(error));
   };
